test(models): keep child boundary proof in wire tests
